refactor: 추천 화면 폴더 구조 변경
